ONIDevice register access simplification
- Remove redundant dev_index argument from ONIDevice.ReadRegister and
ONIDevice.WriteRegister since this is a property of ONIDevice itself
- Lots of other small fixes
- Better looking ONIX logo
- Updated and improved ExampleWorkflows
- Added RealTimeFunctionGenerator for making periodic waveforms that are
more appropriately used with ONIX output devices
- Added some dedicated nodes for visualizing NeuroPixels data (see the
NeuroPixels example workflow)
- Updated and improved example workflows

+++ Bonsai.ONIX.Design/StimulatorEditorDialog.cs
@@ -18,7 +18,7 @@ namespace Bonsai.ONIX.Design
             {
 
                 var context = c.Context;
-                if (context.DeviceTable.TryGetValue(address.Address, out oni.Device dev))
+                if (context.DeviceTable.TryGetValue((uint)address.Address, out oni.Device dev))
                 {
                     device = ONIDeviceFactory.Make((ONIXDevices.ID)dev.ID);
                     if (device != null)
@@ -30,7 +30,7 @@ namespace Bonsai.ONIX.Design
 
                         // Copy Device metadata
                         device.FrameClockHz = c.Context.AcquisitionClockHz;
-                        device.Hub = c.Context.GetHub(device.DeviceAddress.Address);
+                        device.Hub = c.Context.GetHub((uint)device.DeviceAddress.Address);
                     }
                     else
                     {
